feat: cleaning includes and type errors

+++ include/lob/order.hh
@@ -10,19 +10,19 @@ class Order {
     private:
         const unsigned int orderId;
         const unsigned int customerId;
-        const unsigned float amount;
-        const unsigned float pricePerUnit;
+        const float amount;
+        const float pricePerUnit;
         const OrderType type;
         const std::string instrument;
         time_t timestamp;
 
     public:
-        Order(unsigned int orderId, unsigned int customerId, unsigned int amount, unsigned int pricePerUnit, OrderType type, const std::string& instrument);
+        Order(unsigned int orderId, unsigned int customerId, float amount, float pricePerUnit, OrderType type, const std::string& instrument);
 
         unsigned int getOrderId() const;
         unsigned int getCustomerId() const;
-        unsigned int getAmount() const;
-        unsigned int getPricePerUnit() const;
+        float getAmount() const;
+        float getPricePerUnit() const;
         OrderType getType() const;
         std::string getInstrument() const;
         time_t getTimestamp() const;
